:truck: Move fakeData.js to ./helpers/ directory
