+++ models/seeds/003-trainingSeries.js
@@ -1,4 +1,4 @@
-const { createFakeTrainingSeries } = require("../db/fakeData");
+const { createFakeTrainingSeries } = require("../../helpers/fakeData");
 exports.seed = function(knex, Promise) {
   return knex("training_series").then(function() {
     return knex("training_series").insert(createFakeTrainingSeries());
