+++ models/seeds/004-teamMembers.js
@@ -1,4 +1,4 @@
-const { createFakeTeamMembers } = require("../db/fakeData");
+const { createFakeTeamMembers } = require("../../helpers/fakeData");
 exports.seed = function(knex, Promise) {
   return knex("team_members").then(function() {
     return knex("team_members").insert(createFakeTeamMembers());
