+++ models/seeds/005-trainingSeries_teamMembers.js
@@ -1,4 +1,4 @@
-const { createFakeRelationalEntries } = require("../db/fakeData");
+const { createFakeRelationalEntries } = require("../../helpers/fakeData");
 exports.seed = function(knex, Promise) {
   return knex("relational_table").then(function() {
     return knex("relational_table").insert(createFakeRelationalEntries());
