+++ models/seeds/006-messages.js
@@ -1,4 +1,4 @@
-const { createFakeMessages } = require("../db/fakeData");
+const { createFakeMessages } = require("../../helpers/fakeData");
 exports.seed = function(knex, Promise) {
   return knex("messages").then(function() {
     return knex("messages").insert(createFakeMessages());
