+++ models/seeds/007-notifications.js
@@ -1,4 +1,4 @@
-const { createFakeNotifications } = require("../db/fakeData");
+const { createFakeNotifications } = require(".../../helpers/fakeData");
 exports.seed = function(knex, Promise) {
   return knex("notifications").then(function() {
     return knex("notifications").insert(createFakeNotifications());
